requirements.txt
asgiref==3.8.1
Django==5.1.1
django-rest==0.8.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
PyJWT==2.9.0
python-decouple==3.8
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
